fix: Fix relative_file_path method
The method was previously trying to match parts of the Python dotted path
with the file path of the Python module. It failed when the Python object
was declared in an  file.

The method now uses the object's root's file path to compute
the relative path from two levels above it (one for the
file, one more for the package directory).