wip: Refactor how collecting documentation works
A few things first. Since we run tools that collect
documentation in subprocesses, I was previously doing
this in two phases: first, in the on_nav event,
parse all pages to get the autodoc instructions
and group them by handler type, and then
collect all data at once by spawning the subprocess
just once. I was then injecting this data into the
mkdocstrings extension, so that it could use
it when rendering the markdown contents.

This was problematic because we needed to compute
hashes of the autodoc identifiers and selection
options, to make sure we were rendering the right
ones at the right time. Making sure everything
was computed and rendered correctly seemed
quite complex.

But the main reason I did this in two phases,
collecting instructions then running subprocesses,
was a performance reason. Indeed, spawning a new
subprocess (esp. Python)for each autodoc instruction
could make mkdocstrings very slow.

So I wondered: are there other ways to keep good
performance and proceed in a more linear manner,
collecting data and rendering it directly from
within the markdown extension?

Well, the answer is yes! Just have a process
that accepts input on stdin, line by line,
outputing the result on a single line as well,
and keep that process open until site
generation is done :) (then close it).

So that's what I'm doing now: when the markdown
extension hits an autodoc instruction, it gets
the associated handler (which is cached), collects
the data using the handler's collector, and
renders the data using the handler's renderer.

This is so much cleaner <3

Anyway, I just like writing a detailed commit
message sometimes.