fix: Handle `site_url` not being defined
When serving, `site_url` is auto-populated by MkDos.
When building, it is not, and when MkDocstrings tries
to build the URLs mapping, it fails because `page.canonical_url`
is `None`.

This fix catches the exception and silently passes,
but also logs an error if `site_url` is not defined.

Fixes #77.