fix: Use relative links for cross-references
We were previously using canonical URLs to
fix cross-references throughout the built site.
This was not working when the site was served
on a sub-URL, for example an URL including
the version or the language of the docs.

Cross-references URLs are now relative
to each other. This also allow us to get
rid of the check for the configuration
value `site_url`, which is not needed anymore.

Fixes #144.
References #147.