refactor: Don't re-parse the whole subdoc, expose only headings
Previously the autodoc sub-document (HTML with nested Markdown) was being re-integrated into the top-level document by the way of parsing it as XML and inserting it back into the XML tree of the top-level markdown during its rendering.
This gave a lot of side effects that need to be worked around. XML parsing, Markdown double-processing (with only situational workaround possible), etc.

Now, the document will be inserted as fully opaque HTML, so those issues are avoided entirely. But now there are new issues to overcome!

* The ToC extension can no longer see through this subdocument, so no ToC entries can be generated. I thought a lot how to re-add these properly, and there's no easy solution without parsing the HTML. In the end I decided to manually reinsert just the heading elements back into the document (note that detecting headings is two-pronged: in Markdown with a treeprocessor, and in HTML through requiring the use of a specific callback filter). As those headings have no business there, before finishing the document they get removed.

* MkDocs can no longer fix relative paths in this subdocument. That one is much easier: just manually re-add its treeprocessor.