tests: Build test sites into a temp directory
This was causing random test failures because multiple tests concurrently try to delete and re-create the same site dir.
Refactor to use pytest fixtures instead while on it.

Bonus: allow docstrings to not specify Arguments sections etc. But if they are specified, they will still be checked.
Here because it's ridiculous to document fixtures in every test function. Generally because sometimes all args are obvious if it's just an override.