fix: Do minimum work when falling back to re-collecting an object to get its anchor
This change introduces a new attribute on collectors: `fallback_config`.
It is used by the `Handlers.get_anchor` method when collecting an object
to find its anchor. It allows to pass specific options to the collector
to avoid collecting the object members or, for example,
parsing Python docstrings, as it is not useful,
since we only care about the anchor.

Issue #329: https://github.com/mkdocstrings/mkdocstrings/issues/329
PR #330: https://github.com/mkdocstrings/mkdocstrings/issues/330