feat: Register objects' aliases into the inventory
When publishing API docs, maintainers might
inject documentation for an object for only
one of its locations (canonical, public, other
alias, etc.). Without this change, other
doc sites can only cross-reference this
object with the path it was rendered with.
With this change, other doc sites can
cross-reference the object with any of its
aliases.

It means that doc writers do not have to worry
about injecting docs for every object with both
their canonical and public paths: only one path
is fine, and other can cross-ref to it with
any valid alias path.