refactor: Deprecate base handler's `get_anchors` method in favor of `get_aliases` method
First, `get_anchors` accepts a data object as argument. This forces mkdocstrings to (re)collect this data object through the handler's `collect` method just to pass it again to the handler. Why not giving back control to the handler itself? For example, the Python handler would not call `collect` again, but just look into its cached objects collection.

Second, "anchors" in this context doesn't make much sense, so we rename the method `get_aliases`, because it really returns aliases of a given identifier, not "HTML anchors". Plus this makes the deprecation easier (method rename instead of signature change).